Add aarch64 OpenBSD support Notable items: * OpenBSD uses emulated tls so exclude optimized get_thread in   threadLS_bsd_aarch64.s and provide slower implementation for   OpenBSD. * OpenBSD uses unsigned long long for u_int64_t instead of   unsigned long which prevents implicit conversions currently   in use. Adjust to correct datatypes and format specifiers   as needed. * READ_SPECIALREG called from userland on OpenBSD raises   SIGILL/ILL_ILLOPC signal. Disable cpu detection and   processor features and set them to the most compatible values   for OpenBSD.

+++ src/hotspot/cpu/aarch64/macroAssembler_aarch64_log.cpp
@@ -297,7 +297,7 @@ void MacroAssembler::fast_log(FloatRegister vtmp0, FloatRegister vtmp1,
   bind(MAIN);
     fmovs(tmp3, vtmp5);                        // int intB0 = AS_INT_BITS(B);
     mov(tmp5, 0x3FE0);
-    mov(rscratch1, 0xffffe00000000000);
+    mov(rscratch1, (u_int64_t)0xffffe00000000000);
     andr(tmp2, tmp2, tmp1, LSR, 48);           // hiWord & 0x7FF0
     sub(tmp2, tmp2, tmp5);                     // tmp2 = hiWord & 0x7FF0 - 0x3FE0
     scvtfwd(vtmp5, tmp2);                      // vtmp5 = (double)tmp2;
